Split out district boundaries creation into own class
No changes to the algorithm at this point, just moving from ScoutMap
